Tipping Dealer and Odds boosting
The player's odds are now boosted after they tip the dealer

+++ src/main/java/blackjack/Bank.java
@@ -23,7 +23,7 @@ public class Bank {
     }
 
     //Tip the dealer
-    public void tipDealer(double amount) {
+    public void tipDealer(int amount) {
         if(amount <= getCurrency()) {
             System.out.println("Tip dealer: " + amount);
             currency -= amount;
